+++ pylusat/base.py
@@ -3,8 +3,12 @@ from geopandas import GeoDataFrame
 import numpy as np
 import rasterio as rio
 from rasterio.vrt import WarpedVRT
-from rasterio import DatasetReader, MemoryFile
+from rasterio import DatasetReader
+from rasterio.io import MemoryFile
 from rasterio.enums import Resampling
+from shapely.geometry import box
+from shapely.ops import unary_union
+from affine import Affine
 
 
 class GeoDataFrameManager:
@@ -261,4 +265,36 @@ class RasterManager:
                 with MemoryFile() as memfile:
                     rst = memfile.open(**profile)
                     rst.write(data)
-                    return rst
+                    rst.close()
+                    return memfile.open(driver='GTiff')
+
+    def match_extent(self, ref_obj):
+        out_x_res = round(ref_obj.get_affine()[0], 1)
+        out_y_res = round(-ref_obj.get_affine()[4], 1)
+        b1 = box(*self.rast_ds.bounds)
+        b2 = box(*ref_obj.rast_ds.bounds)
+        b_union = unary_union((b1, b2))
+        out_bnd = b_union.envelope
+        left, bottom, right, top = out_bnd.bounds
+        out_width = round((right - left) / out_x_res)
+        out_height = round((top - bottom) / out_y_res)
+        out_transform = Affine(out_x_res, 0, left, 0, -out_y_res, top)
+        out_crs = ref_obj.get_rio_crs()
+        out_nodata = self.rast_nodata
+
+        vrt_options = {
+            'resampling': Resampling.nearest,
+            'crs': out_crs,
+            'transform': out_transform,
+            'height': out_height,
+            'width': out_width,
+            'nodata': out_nodata
+        }
+        out_vrt = WarpedVRT(self.rast_ds, **vrt_options)
+
+        with MemoryFile() as memfile:
+            rst = memfile.open(driver='GTiff', count=1,
+                               dtype='int32', **vrt_options)
+            rst.write(out_vrt.read(1), indexes=1)
+            rst.close()
+            return memfile.open(driver="GTiff")