+++ pylusat/tests/test_geotools.py
@@ -1,5 +1,5 @@
 import geopandas as gpd
-from pylusat.geotools import spatial_join, select_by_location
+from pylusat.geotools import spatial_join, select_by_location, combine
 from pylusat.datasets import get_path
 import pytest
 
@@ -14,6 +14,11 @@ def schools_gdf():
     return gpd.read_file(get_path("schools"))
 
 
+@pytest.fixture
+def habitat_tif():
+    return get_path("habitat")
+
+
 def test_spatial_join(acs2016_gdf, schools_gdf):
     result = spatial_join(acs2016_gdf, schools_gdf,
                           cols_agg={'ENROLLMENT': ['sum']})
@@ -23,3 +28,9 @@ def test_spatial_join(acs2016_gdf, schools_gdf):
 def test_select_by_location(acs2016_gdf, schools_gdf):
     result = select_by_location(acs2016_gdf, schools_gdf)
     assert len(result) == 75
+
+
+def test_combine(habitat_tif):
+    rast_obj, attr = combine(habitat_tif, habitat_tif)
+    assert len(attr) == 29
+    assert attr['count'][attr.value == 1][0] == 618688